backend/requirements.txt
chromadb>=0.4.0
Pillow>=9.0.0
requests>=2.28.0
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0 
langchain_groq 
langchain_community 
langchain 
langchain_core 
langchain_chroma 
faster_whisper
uvicorn 
fastapi 
groq 
numpy 
torch 